requirements.txt
click==8.1.3
matplotlib==3.5.2
numpy==1.22.4
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.2.0
PyYAML==6.0
scikit_learn==1.1.2
scipy==1.8.1
seaborn==0.11.2
setuptools==62.5.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
